requirements.txt
Django==4.1.3
djangorestframework==3.15.1
gunicorn==22.0.0
whitenoise==6.7.0
dj-database-url==2.2.0
django-cors-headers==4.4.0
psycopg2-binary==2.9.11
Pillow==12.1.0
djangorestframework-simplejwt
django-filter==25.2
